refactor: extract sharing feature folder
Move invite-links, share-dialog, space-share-dialog,
sidebar-collaboration, and invite screen into
src/app/features/sharing/.
