+++ src/app/features/sharing/widgets/space-share-dialog.tsx
