+++ src/app/routes/doc.$id.index.tsx
@@ -68,7 +68,7 @@ import { SidebarViewLinks } from "@/components/sidebar-view-links"
 import { SidebarFileMenu } from "@/components/sidebar-file-menu"
 import { SidebarEditMenu } from "@/components/sidebar-edit-menu"
 import { SidebarFormatMenu } from "@/components/sidebar-format-menu"
-import { SidebarCollaboration } from "@/components/sidebar-collaboration"
+import { SidebarCollaboration } from "@/app/features/sharing"
 import { SidebarAssets, type SidebarAsset } from "@/components/sidebar-assets"
 import { ThemeToggle, useTheme } from "@/app/components/appearance"
 import {
