+++ src/app/routes/spaces.$spaceId.doc.$id.index.tsx
@@ -70,7 +70,7 @@ import { SidebarViewLinks } from "@/components/sidebar-view-links"
 import { SidebarFileMenu } from "@/components/sidebar-file-menu"
 import { SidebarEditMenu } from "@/components/sidebar-edit-menu"
 import { SidebarFormatMenu } from "@/components/sidebar-format-menu"
-import { SidebarCollaboration } from "@/components/sidebar-collaboration"
+import { SidebarCollaboration } from "@/app/features/sharing"
 import { SidebarAssets, type SidebarAsset } from "@/components/sidebar-assets"
 import { ThemeToggle, useTheme } from "@/app/components/appearance"
 import {
