+++ src/app/routes/spaces.$spaceId.settings.tsx
@@ -7,7 +7,7 @@ import { ArrowLeft, Loader2, Trash2, Upload, UserRoundPlus } from "lucide-react"
 import Cropper from "react-easy-crop"
 import { Space, UserAccount } from "@/schema"
 import { permanentlyDeleteSpace } from "@/lib/spaces"
-import { SpaceShareDialog } from "@/components/space-share-dialog"
+import { SpaceShareDialog } from "@/app/features/sharing"
 import { SpaceInitials } from "@/components/space-selector"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
