+++ src/components/sidebar-document-list.tsx
@@ -71,7 +71,7 @@ import {
 import { useFolderStore, FolderRow } from "@/components/folder"
 import { getPresentationMode } from "@/lib/presentation"
 import { exportDocument, type ExportAsset } from "@/lib/export"
-import { ShareDialog } from "@/components/share-dialog"
+import { ShareDialog } from "@/app/features/sharing"
 import { MoveToFolderDialog } from "@/components/move-to-folder-dialog"
 import { MoveToSpaceDialog } from "@/components/move-to-space-dialog"
 import { ConfirmDialog } from "@/components/ui/confirm-dialog"
